Deleted all benchmarking/exploratory files/deps
