+++ .gitignore
@@ -2,3 +2,4 @@
 /deps
 erl_crash.dump
 *.ez
+*.beam
