--- bench/snapshots/2015-05-15T16:26:25.snapshot
@@ -1,2 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
