--- bench/snapshots/2015-05-15T17:11:00.snapshot
@@ -1,2 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
