--- bench/snapshots/2015-05-15T18:07:11.snapshot
@@ -1,6 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
-NaturalSortBench	natural sort using explicit recursion to break up the strings		10000	2985834
-NaturalSortBench	natural sort using Regex.scan to break up the strings		5000	2105713
-RegexFormatterBench	Original format function recursing through string		50000	1890453
-RegexFormatterBench	New format function using scan		50000	2134616
