--- bench/snapshots/2015-05-16T08:55:45.snapshot
@@ -1,4 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
-RegexFormatterBench	Original format function recursing through string		50000	1980516
-RegexFormatterBench	New format function using scan		50000	2150618
