--- bench/snapshots/2015-05-16T09:10:20.snapshot
@@ -1,7 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
-NaturalSortBench	second natural sort using Regex.scan to break up the strings		50	1988418
-NaturalSortBench	natural sort using explicit recursion to break up the strings		50	2012817
-NaturalSortBench	natural sort using Regex.scan to break up the strings		50	1909858
-RegexFormatterBench	Original format function recursing through string		50000	1916308
-RegexFormatterBench	New format function using scan		50000	2165338
