--- bench/snapshots/2015-05-16T09:17:57.snapshot
@@ -1,7 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
-NaturalSortBench	second natural sort using Regex.scan to break up the strings		50	2096318
-NaturalSortBench	natural sort using explicit recursion to break up the strings		50	2089018
-NaturalSortBench	natural sort using Regex.scan to break up the strings		50	2002039
-RegexFormatterBench	Original format function recursing through string		50000	1992971
-RegexFormatterBench	New format function using scan		50000	2201550
