--- bench/snapshots/2015-05-16T09:19:04.snapshot
@@ -1,8 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
-NaturalSortBench	second natural sort using Regex.scan to break up the strings		50	2042285
-NaturalSortBench	natural sort using explicit recursion to break up the strings		50	2061336
-NaturalSortBench	natural sort using Regex.scan to break up the strings		50	1941694
-RegexFormatterBench	Original format function recursing through string		50000	1917756
-RegexFormatterBench	Optimised New format function using scan		50000	2164670
-RegexFormatterBench	New format function using scan		50000	2126692
