--- bench/snapshots/2015-05-16T09:21:30.snapshot
@@ -1,8 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
-NaturalSortBench	second natural sort using Regex.scan to break up the strings		50	2185366
-NaturalSortBench	natural sort using explicit recursion to break up the strings		50	2134676
-NaturalSortBench	natural sort using Regex.scan to break up the strings		50	2048528
-RegexFormatterBench	Original format function recursing through string		50000	1991020
-RegexFormatterBench	Optimised New format function using scan		50000	2223780
-RegexFormatterBench	New format function using scan		50000	2150025
