--- bench/snapshots/2015-05-16T09:29:48.snapshot
@@ -1,8 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
-NaturalSortBench	second natural sort using Regex.scan to break up the strings		100	1959273
-NaturalSortBench	natural sort using explicit recursion to break up the strings		50	2120533
-NaturalSortBench	natural sort using Regex.scan to break up the strings		50	2029601
-RegexFormatterBench	Original format function recursing through string		50000	1981467
-RegexFormatterBench	Optimised New format function using scan		100000	2247882
-RegexFormatterBench	New format function using scan		50000	2199814
