--- bench/snapshots/2015-05-16T09:38:15.snapshot
@@ -1,8 +1,0 @@
-duration:1.0;mem stats:false;sys mem stats:false
-module;test;tags;iterations;elapsed
-NaturalSortBench	second natural sort using Regex.scan to break up the strings		100	2007589
-NaturalSortBench	natural sort using explicit recursion to break up the strings		50	2038105
-NaturalSortBench	natural sort using Regex.scan to break up the strings		50	1971180
-RegexFormatterBench	Original format function recursing through string		50000	1893315
-RegexFormatterBench	Optimised New format function using scan		100000	2271664
-RegexFormatterBench	New format function using scan		50000	2121367
