+++ mix.exs
@@ -16,9 +16,6 @@ defmodule NaturalSort.Mixfile do
   end
 
   defp deps do
-    [
-      {:benchfella, "~> 0.2"},
-      {:faker, "~> 0.5"}
-    ]
+    []
   end
 end
